remove update and create listener to use commands in a single file. add the prefix property to reuse it
