--- src/main/java/com/swamigallardo/mandalorian/MessageCreateListener.java
@@ -1,19 +1,0 @@
-package com.swamigallardo.mandalorian;
-
-import discord4j.core.event.domain.message.MessageCreateEvent;
-import org.springframework.stereotype.Service;
-import reactor.core.publisher.Mono;
-
-@Service
-public class MessageCreateListener extends MessageListener implements EventListener<MessageCreateEvent> {
-
-    @Override
-    public Class<MessageCreateEvent> getEventType() {
-        return MessageCreateEvent.class;
-    }
-
-    @Override
-    public Mono<Void> execute(MessageCreateEvent event) {
-        return processCommand(event.getMessage());
-    }
-}
