--- src/main/java/com/swamigallardo/mandalorian/MessageListener.java
@@ -1,17 +1,0 @@
-package com.swamigallardo.mandalorian;
-
-import discord4j.core.object.entity.Message;
-import reactor.core.publisher.Mono;
-public abstract class MessageListener{
-
-    public Mono<Void> processCommand(Message eventMessage) {
-
-        return Mono.just(eventMessage)
-                .filter(message -> message.getAuthor().map(user -> !user.isBot()).orElse(false))
-                .filter(message -> message.getContent().equalsIgnoreCase("!ping"))
-                .flatMap(Message::getChannel)
-                .flatMap(channel -> channel.createMessage("pong"))
-                .then();
-    }
-
-}
