--- src/main/java/com/swamigallardo/mandalorian/MessageUpdateListener.java
@@ -1,24 +1,0 @@
-package com.swamigallardo.mandalorian;
-
-import com.swamigallardo.mandalorian.EventListener;
-import com.swamigallardo.mandalorian.MessageListener;
-import discord4j.core.event.domain.message.MessageUpdateEvent;
-import org.springframework.stereotype.Service;
-import reactor.core.publisher.Mono;
-
-@Service
-public class MessageUpdateListener extends MessageListener implements EventListener<MessageUpdateEvent> {
-
-    @Override
-    public Class<MessageUpdateEvent> getEventType() {
-        return MessageUpdateEvent.class;
-    }
-
-    @Override
-    public Mono<Void> execute(MessageUpdateEvent event) {
-        return Mono.just(event)
-                .filter(MessageUpdateEvent::isContentChanged)
-                .flatMap(MessageUpdateEvent::getMessage)
-                .flatMap(super::processCommand);
-    }
-}
