+++ src/main/resources/application.properties
@@ -1,1 +1,2 @@
 mandalorian.ping.up=Mandalorian is up!
+prefix=!
